1. setup poetry 2. setup the chains class 3. still have an error
